chore: extract bufferToSha1 to it's own utils file
